requirements.txt
Flask==2.3.3
pyvmomi==8.0.2.0.1
python-dotenv==1.0.0
Flask-WTF==1.1.1
pyVim==3.0.3
requests==2.31.0
